Elf32: pass endianness from Ehdr to Shdr and Phdr
Section and program headers were parsed as big-endian regardless of real
endianness, when whole ELF was deserialized using from_bytes method.
This was because endianness indicator was not passed from Ehdr to them.
Therefore, Ehdr was parsed correctly and other headers were also
correctly dissected, but next, sections itself were not parsed, because
of wrong endianness of section headers.

This should solve issue #2.