sh_type: use proper endianness
This field is quite special, becuase it is the only enum in Shdr and as
an enum it has to be reversed in order after serialization (not
automatically, by passing right order as boolean) and because it was
lone, it was missed till now.

Furthermore there exists a hack for unknown values, so it can be either
enum or integer. And in the latter case it should not be reversed.